Switch to Wikimedia Commons as default site
Commons is intended as a repository for freely-usable media, so it makes more
sense for it to be the default, and most images on Wikimedia sites (like English
Wikipedia) are also available there.

The functionality for specifying alternate sites is left in case users want to
download fair-use media, which is not available on Commons, as well as for the
possible future integration with non-Wikimedia MediaWiki sites (like Fandom/Wikia).