Prevent unintended behavior when combining -o and -a
When batch downloading files with -a, specifying an output name
with -o will result in only the first file being downloaded. If
-f is also given, then only the last file is kept (with the previous
ones being overwritten). This is not ideal behavior, so the ability
to specify an output filename has been disabled when running wikiget
in batch mode.

In the future, -o could be used as a template string or prefix when
in batch mode, along the lines of what youtube-dl does.