Add parallel download option in batch mode
Number of download threads can be set with new -j
option. Unfortunately, it's not that much faster
than downloading in serial, since the API calls
made before the downloads actually start are not
(and ideally should not be) parallelized. Still,
for large batches, it saves a bit of time.

Known issue: due to the download threads writing
to the log asynchronously, the messages get
jumbled up. This will be fixed eventually.